dailyChallenge/requirements.txt
discord.py
geoguessr_async
pycountry
dotenv